mcp-server/requirements.txt
fastmcp>=0.4.0
httpx>=0.25.0
pynacl>=1.5.0
